requirements_new.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
av==12.0.0
backcall==0.2.0
backports.tarfile==1.1.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docutils==0.20.1
environs==9.5.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
fastdtw==0.3.4
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.46.0
fsspec==2024.3.1
fvcore==0.1.5.post20221221
gradio==4.27.0
gradio-client==0.15.1
grpcio==1.53.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==7.1.0
importlib-resources==6.4.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
keyring==25.1.0
kiwisolver==1.4.5
lazy-loader==0.3
librosa==0.10.1
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.1.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==3.0.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
mysql-connector-python==8.0.17
nest-asyncio==1.6.0
networkx==3.1
nh3==0.2.17
numba==0.58.1
numpy==1.24.4
opencv-python==4.8.1.78
orjson==3.10.1
packaging==23.2
pandas==2.0.3
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pkginfo==1.10.0
pkgutil-resolve-name==1.3.10
platformdirs==4.0.0
pooch==1.8.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.11
pycparser==2.21
pydantic==2.7.0
pydantic-core==2.18.1
PyDejavu==0.1.3
pydub==0.23.1
pygments==2.17.2
pymilvus==2.2.11
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytorchvideo==0.1.5
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
readme-renderer==43.0
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.1
scenedetect==0.6.2
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
towhee==1.1.3
towhee.models==1.1.3
tqdm==4.66.1
traitlets==5.14.3
twine==5.0.0
typer==0.12.3
typing-extensions==4.8.0
tzdata==2024.1
ujson==5.4.0
urllib3==2.1.0
uvicorn==0.29.0
wcwidth==0.2.13
websockets==11.0.3
yacs==0.1.8
zipp==3.18.1
